Update dotnet.yml
The error message was because the dotnet restore command is being run in a directory that contains more than one project or solution file. The dotnet restore command needs to know which project or solution file to use, and if there’s more than one, it doesn’t know which one to pick.

You can specify the project or solution file directly in the command. For example, if your project file is named MyProject.csproj